Don’t import Error and ErrorKind in lib.rs
Currently, if a user tries to access genpdf::Error, they will get an
error message that states that Error is private.  This can be
misleading, because the import in lib.rs is private, while the struct
itself (genpdf::error::Error) is of course public.

This patch tries to improve this error message by removing the Error and
ErrorKind imports from lib.rs.