Separate font loading and font caching
With this cache, we separate the font loading from the font caching:  We
add the FontData::load and fonts::from_files functions to load fonts and
font families.  And we replace the load_font* methods in FontCache and
Document with add_font*.

This gives us more flexibility and allows the user to manually load
fonts, e. g. from embedded data or from the network.