Check font validity in FontData::new
We can only handle fonts that are scalable.  Previously, we performed
that check in Font::new.  This meant that all methods that add fonts to
the cache had to return a Result.  With this patch, we move the font
check into the FontData::new constructor.  This allows us to have a
single point of error for font loading and to make the the methods that
add fonts to the cache infallible.