Add support for built-in fonts
This patch adds support for the three built-in font families, Helvetica,
Times and Courier.  While we still need font data of a metrically
identical font to query the gliph metrics, we don’t have to embed these
fonts into the PDF file because the PDF viewer is expected to support
them.  Currently, the built-in fonts are limited to ASCII characters
because of an issue in printpdf, see this upstream issue:
	https://github.com/fschutt/printpdf/issues/30