Add StyledCow<'s> for a styled Cow<'s, str>
This patch adds the StyledCow<'s> struct that contains a Cow<'s, str>
with a Style annotation.  This allows us to write functions that can
return both references to an existing styled string or create a new
styled string.  This is, for example, required for hyphenation as we
want to be able to append a hyphen to hyphenated words.