Add the PageDecorator trait to customize pages
This patch adds the PageDecorator trait that can be used to customize
all pages after they have been created by Document and before they are
filled with content.  We also add the SimplePageDecorator struct as a
basic implementation that supports margins and a page header.

We remove the Document::add_margins method as it is now redundant.