Keep track of string length changes during wrapping
This patch fixes a bug that was introduced when we added hyphenation
support:

A Paragraph has to keep track of the data it already rendered.  We used
to do this by calculating the length of the printed strings and using
this length as an offset from the input data in self.text.  This fails
if the word wrapper modifies the input data.  This is the case when a
word is split and a hyphen is added:  The printed data is longer than
the input data, so we would skip more of the input data than we actually
printed.

To fix this issue, we add an additional value to the item type of the
Wrapper iterator:  the number of characters that was added or removed
from the input string.

We also convert the input data into words on the first call of render.
This makes it easier to keep track of the rendering progress:  We just
drop everything that we have rendered.